Add files via upload
Each operation in the SAP-1’s instruction set have been mapped to 4 bits. These 4 bits make up the first 4 of the total 8 bits that are stored in each RAM location. The last 4 bits specify the operand which the operation will act on. Example – 000 0001 will load (0000) the operand 1 (0001) to the accumulator.